requirements-db.txt
# Database dependencies for AlphaPulse

# PostgreSQL and TimescaleDB
asyncpg>=0.27.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
alembic>=1.10.0

# Redis
redis>=4.5.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
bcrypt>=4.0.0